fix(runtime,zia,codegen): Windows completeness — string refcount bugs, linker, platform gaps
- Fix Zia string concat double-free: rt_str_concat consumes both operands
  internally, but the deferred release system also emitted releases for
  them. Add consumeDeferred() calls after kStringConcat to prevent this.

- Fix Zia entity field store missing retain/release: emitFieldStore() was
  not retaining strings stored into entity fields, causing dangling
  pointers after the caller's deferred release fires. Add retain-new and
  release-old logic for Str fields (BUG-ADV-002).

- Fix Windows linker missing vipergui: add vipergui to backendLibs and
  extend findLibArchive() to search src/lib/ subdirectories.

- Fix lseek 32-bit mapping: change _lseek to _lseeki64 in rt_platform.h
  for >2GB file support; update rt_io.c to use int64_t instead of off_t.

- Fix rt_dns_local_addrs() on Windows: replace gethostname/getaddrinfo
  with GetAdaptersAddresses for complete interface enumeration.

- Fix WM_SIZE handler: don't update framebuffer dimensions on resize
  (only update logical StretchBlt dimensions) to prevent heap overflow.

- Fix PkgDeflate.cpp MSVC C4334 warning (32-bit shift result).

- Uncomment viperide in build_demos.cmd, update graphics docs to reflect
  complete Linux/Windows backend status.

+++ src/tools/common/packaging/PkgDeflate.cpp
@@ -243,7 +243,7 @@ struct HuffmanTree {
             maxLen = 1;
 
         tableBits = maxLen;
-        tableSize = 1U << tableBits;
+        tableSize = static_cast<size_t>(1) << tableBits;
         maxCode = numCodes;
 
         symbols = static_cast<uint16_t *>(
